refactor: 객체를 생성하는 로직은 Controller 패키지로 분리
