+++ src/main/java/laddergame/controller/LadderGame.java
@@ -2,7 +2,6 @@ package laddergame.controller;
 
 import laddergame.controller.rule.Rule;
 import laddergame.domain.Ladder;
-import laddergame.domain.LadderGenerator;
 import laddergame.domain.Tags;
 import laddergame.util.OutputView;
 
@@ -16,8 +15,8 @@ public class LadderGame {
 
     public LadderGame(Tags members, Rule rule) {
         this.members = members;
-        this.prizes = Maker.makePrizes(members.size());
-        this.ladder = LadderGenerator.generateLadder(Maker.makeHeight(), members.size(), rule);
+        this.prizes = GamePreparer.makePrizes(members.size());
+        this.ladder = LadderGenerator.generateLadder(GamePreparer.makeHeight(), members.size(), rule);
     }
 
     public LadderGame(Tags members, Tags prizes, Rule rule, int height) {
