+++ src/test/java/laddergame/domain/LadderTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.controller.LadderGenerator;
 import laddergame.controller.rule.AlwaysCreateRule;
 import org.junit.jupiter.api.Test;
